refactoring the architechture
