+++ src/controllers/Article/index.js
@@ -1,0 +1,5 @@
+module.exports = {
+  getAllArticles: require("./getAllArticles"),
+  createArticle: require("./createArticle"),
+  queryArticle: require("./queryArticle"),
+}
