+++ src/controllers/Duck/getDuck.js
@@ -6,3 +6,4 @@ module.exports = (req, res) => {
     {},
   )));
 };
+
