+++ src/controllers/Duck/index.js
@@ -1,0 +1,6 @@
+
+module.exports = {
+  getDuck: require("./getDuck"),
+  getDuckById: require("./getDuckById"),
+  getZoomedDuck: require("./getZoomedDuck"),
+}
