+++ src/controllers/Events/index.js
@@ -1,0 +1,9 @@
+module.exports = {
+    createEvent : require('./createEvent'),
+    deleteEvent : require('./deleteEvent'),
+    editEvent : require('./editEvent'),
+    listEvents : require('./listEvents'),
+    queryEvent : require('./queryEvent'),
+    rsvpEvent : require('./rsvpEvent'),
+    unrsvpEvent : require('./unrsvpEvent'),
+}
