+++ src/controllers/Joke/getAllJokes.js
@@ -1,0 +1,7 @@
+const jokeLib = require("./jokeFinder");
+
+
+module.exports = (req, res) => {
+    res.set("Access-Control-Allow-Origin", "*");
+    return res.json(jokeLib.getAllJokes());
+};
