+++ src/controllers/Joke/getJokeCount.js
@@ -1,0 +1,6 @@
+const jokeLib = require("./jokeFinder");
+
+module.exports.get_joke_count = (req, res) => {
+    res.set("Access-Control-Allow-Origin", "*");
+    return res.send(jokeLib.getJokeCount().toString());
+};
