+++ src/controllers/Joke/getRandomJoke.js
@@ -1,0 +1,6 @@
+const jokeLib = require("./jokeFinder");
+
+module.exports = (req, res) => {
+    res.set("Access-Control-Allow-Origin", "*");
+    return res.send(jokeLib.getRandomJoke());
+};
