+++ src/controllers/Joke/index.js
@@ -1,0 +1,6 @@
+module.exports = {
+    getAllJokes : require("./getAllJokes"),
+    getJokeById : require("./getJokeById"),
+    getJokeCount : require("./getJokeCount"),
+    getRandomJoke : require("./getRandomJoke"),
+}
