--- src/routes/Joke/all.js
@@ -1,6 +1,0 @@
-const jokeLib = require("./jokeFinder");
-
-module.exports = (req, res) => {
-  res.set("Access-Control-Allow-Origin", "*");
-  return res.json(jokeLib.getAllJokes());
-};
