+++ src/routes/articleRoutes.js
@@ -1,0 +1,12 @@
+
+const express = require(express);
+const router = express.Router();
+const  articleControllers = require("../controllers/Article");
+
+
+router.get('/', articleControllers.getAllArticles);
+router.get('/create', articleControllers.createArticle);
+router.get('/:id', articleControllers.queryArticle);
+
+
+module.exports = router;
