+++ src/routes/eventRoutes.js
@@ -1,0 +1,8 @@
+
+const express = require(express);
+const router = express.Router();
+const  eventControllers = require("../controllers/Events");
+
+
+module.exports = router;
+
