+++ src/routes/userRoutes.js
@@ -1,0 +1,10 @@
+
+const express = require(express);
+const router = express.Router();
+const  = require(../controllers/);
+
+
+
+module.exports = router;
+
+ duckRoutes.js eventRoutes.js articleRoutes.js jokeRoutes.js
